Use functional tests (fixes #5, #10)

+++ mapentity/helpers.py
@@ -121,6 +121,9 @@ def is_file_newer(path, date_update, delete_empty=True):
     if not os.path.exists(path):
         return False
 
+    if date_update is None:
+        return False
+
     if os.path.getsize(path) == 0:
         if delete_empty:
             os.remove(path)
@@ -253,9 +256,6 @@ def extract_attributes_html(url):
     With this, we save a lot of efforts, since we do have to build specific Appy.pod
     templates for each model.
     """
-    if getattr(settings, 'TEST', False):
-        return '<p>Mock</p>'  # TODO: better run in LiveServerTestCase instead !
-
     r = requests.get(url)
     if r.status_code != 200:
         raise ValueError('Could not reach %s' % url)
